scraper/requirements.txt
playwright
python-dateutil
markdown
langchain-community
pandas